Conflicts:
	robot/camera.py
	robot/checkdaemon.sh
	robot/sensor.py
	robot/tempreader.sh
	robot/test.py
	robot/test/x.py
	robot/xsend.py
	robot/xtalk.py

+++ robot/sensor.py
@@ -23,7 +23,7 @@ def read_brightness():
         ser.write("LIGHT")
         time.sleep(0.1)
         t = ser.readline()
-        v = float(t)
+        v = int(t)
         time.sleep(0.1)
         ser.close()
         return 1, v
